requirements.txt
alembic==1.8.1
attrs==22.1.0
autopage==0.5.1
catboost==1.1.1
category-encoders==2.5.1.post0
certifi==2022.9.24
charset-normalizer==2.1.1
cliff==4.0.0
cloudpickle==2.2.0
cmaes==0.8.2
cmd2==2.4.2
colorlog==6.7.0
colour==0.1.5
contourpy==1.0.6
cycler==0.11.0
deap==1.3.3
dtreeviz==1.4.0
exceptiongroup==1.0.1
fonttools==4.38.0
graphviz==0.20.1
greenlet==2.0.1
idna==3.4
importlib-metadata==4.13.0
iniconfig==1.1.1
joblib==1.2.0
kiwisolver==1.4.4
lightgbm==3.3.3
llvmlite==0.39.1
Mako==1.2.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mljar-supervised==0.11.3
numba==0.56.4
numpy==1.23.4
optuna==3.0.3
packaging==21.3
pandas==1.5.1
patsy==0.5.3
pbr==5.11.0
Pillow==9.3.0
plotly==5.11.0
pluggy==1.0.0
prettytable==3.5.0
pyparsing==3.0.9
pyperclip==1.8.2
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
scikit-learn==1.1.3
scikit-plot==0.3.7
scipy==1.8.1
seaborn==0.12.1
shap==0.41.0
six==1.16.0
slicer==0.0.7
SQLAlchemy==1.4.43
statsmodels==0.13.5
stevedore==4.1.0
stopit==1.1.2
tabulate==0.9.0
tenacity==8.1.0
threadpoolctl==3.1.0
tomli==2.0.1
TPOT==0.11.7
tqdm==4.64.1
typing_extensions==4.4.0
update-checker==0.18.0
urllib3==1.26.12
wcwidth==0.2.5
wordcloud==1.8.2.2
xgboost==1.7.1
zipp==3.10.0
